Business/gRPC/client/requirements.txt
grpcio
grpcio-tools
mysql-connector-python
locust